requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
scipy==1.10.1
matplotlib==3.7.2
seaborn==0.12.2

# ML and anomaly detection
tensorflow==2.12.0
pytorch==2.0.1
pyod==1.1.1
alibi-detect==0.11.2

# Data validation and quality
great-expectations==0.16.8
pandera==0.15.1
evidently==0.3.3

# API and backend
fastapi==0.98.0
uvicorn==0.22.0
pydantic==1.10.9
python-multipart==0.0.6
sqlalchemy==2.0.17
alembic==1.11.1

# UI
streamlit==1.23.1
plotly==5.15.0
dash==2.11.1

# Utilities
joblib==1.2.0
pyyaml==6.0
tqdm==4.65.0
dill==0.3.6
typing-extensions==4.6.3

# Testing
pytest==7.3.1
pytest-cov==4.1.0
hypothesis==6.81.0

# Documentation
sphinx==7.1.0
sphinx-rtd-theme==1.2.2

# Version control
dvc==3.5.1 